Fixed issue that broke our custom types
