+++ lua.cpp
@@ -1,27 +1,6 @@
 
 #include "lua.h"
 
-// These 2 macros helps us in constructing general metamethods.
-// We can use "lua" as a "Lua" pointer and arg1, arg2, ..., arg5 as Variants objects
-// Check examples in createVector2Metatable
-#define LUA_LAMBDA_TEMPLATE(_f_) \
- [](lua_State* inner_state) -> int {                      \
-     lua_pushstring(inner_state,"__Lua");                 \
-     lua_rawget(inner_state,LUA_REGISTRYINDEX);           \
-     Lua* lua = (Lua*) lua_touserdata(inner_state,-1);;   \
-     lua_pop(inner_state,1);                              \
-     Variant arg1 = lua->getVariant(1);          					\
-     Variant arg2 = lua->getVariant(2);          					\
-     Variant arg3 = lua->getVariant(3);          					\
-     Variant arg4 = lua->getVariant(4);          					\
-     Variant arg5 = lua->getVariant(5);          					\
-     _f_                                         					\
-}
-#define LUA_METAMETHOD_TEMPLATE( lua_state , metatable_index , metamethod_name , _f_ )\
-lua_pushstring(lua_state,metamethod_name); \
-lua_pushcfunction(lua_state,LUA_LAMBDA_TEMPLATE( _f_ )); \
-lua_settable(lua_state,metatable_index-2);
-
 Lua::Lua(){
 	// Createing lua state instance
 	state = luaL_newstate();
@@ -37,6 +16,7 @@ Lua::Lua(){
 	createVector2Metatable(); // "mt_Vector2"
 	createVector3Metatable(); // "mt_Vector3"
 	createColorMetatable(); // "mt_Color"
+    createRect2Metatable(); // "mt_Rect2"
 
 	// Exposing basic types constructors
 	exposeConstructors();
@@ -193,7 +173,9 @@ Variant Lua::callFunction( String function_name, Array args) {
         handleError( ret );
         return 0;
     }
-    return getVariant(1);
+    Variant toReturn = getVariant(1);
+    lua_pop(state, 1);
+    return toReturn;
 }
 
 bool Lua::luaFunctionExists(String function_name){
@@ -246,7 +228,7 @@ bool Lua::pushVariant(Variant var) {
             lua_pushinteger(state, (int64_t)var);
             break;
         case Variant::Type::FLOAT:
-            lua_pushnumber(state,var.operator double());
+            lua_pushnumber(state,var.operator float());
             break;
         case Variant::Type::BOOL:
             lua_pushboolean(state, (bool)var);
@@ -298,7 +280,13 @@ bool Lua::pushVariant(Variant var) {
             void* userdata = (Variant*)lua_newuserdata( state , sizeof(Variant) );
             memcpy( userdata , (void*)&var , sizeof(Variant) );
             luaL_setmetatable(state,"mt_Color");
-            break;    
+            break;
+        }
+        case Variant::Type::RECT2: {
+            void* userdata = (Variant*)lua_newuserdata( state , sizeof(Variant) );
+            memcpy( userdata , (void*)&var , sizeof(Variant) );
+            luaL_setmetatable(state,"mt_Rect2");
+            break;     
         }
         default:
             print_error( vformat("Can't pass Variants of type \"%s\" to Lua." , Variant::get_type_name( var.get_type() ) ) );
@@ -321,7 +309,7 @@ bool Lua::pushGlobalVariant(Variant var, String name) {
 Variant Lua::pullVariant(String name){
     int type = lua_getglobal(state, name.ascii().get_data());
     Variant val = getVariant(1);
-    
+    lua_pop(state, 1);
     return val;
 }
 
@@ -358,217 +346,7 @@ Variant Lua::getVariant(int index) {
         default:
             result = Variant();
     }
-    lua_pop(state, 1);
     return result;
-}
-
-
-void Lua::exposeConstructors( ){
-	
-	lua_pushcfunction(state,LUA_LAMBDA_TEMPLATE({
-        int argc = lua_gettop(inner_state);
-        if( argc == 0 ){
-            lua->pushVariant( Vector2() );
-        } else {
-		    lua->pushVariant( Vector2( arg1.operator double() , arg2.operator double() ) );
-        }
-		return 1;
-	}));
-	lua_setglobal(state, "Vector2" );
-
-	lua_pushcfunction(state,LUA_LAMBDA_TEMPLATE({
-        int argc = lua_gettop(inner_state);
-        if( argc == 0 ){
-            lua->pushVariant( Vector3() );
-        } else {
-		    lua->pushVariant( Vector3( arg1.operator double() , arg2.operator double() , arg3.operator double() ) );
-        }
-        return 1;
-	}));
-	lua_setglobal(state, "Vector3" );
-
-	lua_pushcfunction(state,LUA_LAMBDA_TEMPLATE({
-        int argc = lua_gettop(inner_state);
-        if( argc == 3 ){
-		    lua->pushVariant( Color( arg1.operator double() , arg2.operator double() , arg3.operator double() ) );
-        } else if ( argc == 4 ) {
-		    lua->pushVariant( Color( arg1.operator double() , arg2.operator double() , arg3.operator double() , arg4.operator double() ) );
-        } else {
-            lua->pushVariant( Color() );
-        }
-		return 1;
-	}));
-	lua_setglobal(state, "Color" );
-}
-
-// Create metatable for Vector2 and saves it at LUA_REGISTRYINDEX with name "mt_Vector2"
-void Lua::createVector2Metatable( ){
-    luaL_newmetatable( state , "mt_Vector2" );
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__index" , {
-		lua->pushVariant( arg1.get( arg2 ) );
-		return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__newindex" , {
-		// We can't use arg1 here because we need to reference the userdata
-		((Variant*)lua_touserdata(inner_state,1))->set( arg2 , arg3 );
-		return 0;
-	});	
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__add" , {
-		lua->pushVariant( arg1.operator Vector2() + arg2.operator Vector2() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__sub" , {
-		lua->pushVariant( arg1.operator Vector2() - arg2.operator Vector2() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__mul" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::VECTOR2:
-				lua->pushVariant( arg1.operator Vector2() * arg2.operator Vector2() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Vector2() * arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__div" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::VECTOR2:
-				lua->pushVariant( arg1.operator Vector2() / arg2.operator Vector2() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Vector2() / arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-    lua_pop(state,1); // Stack is now unmodified
-}
-
-// Create metatable for Vector3 and saves it at LUA_REGISTRYINDEX with name "mt_Vector3"
-void Lua::createVector3Metatable( ){
-
-    luaL_newmetatable( state , "mt_Vector3" );
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__index" , {
-		lua->pushVariant( arg1.get( arg2 ) );
-		return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__newindex" , {
-		// We can't use arg1 here because we need to reference the userdata
-		((Variant*)lua_touserdata(inner_state,1))->set( arg2 , arg3 );
-		return 0;
-	});	
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__add" , {
-		lua->pushVariant( arg1.operator Vector3() + arg2.operator Vector3() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__sub" , {
-		lua->pushVariant( arg1.operator Vector3() - arg2.operator Vector3() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__mul" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::VECTOR3:
-				lua->pushVariant( arg1.operator Vector3() * arg2.operator Vector3() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Vector3() * arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__div" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::VECTOR3:
-				lua->pushVariant( arg1.operator Vector3() / arg2.operator Vector3() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Vector3() / arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-    lua_pop(state,1); // Stack is now unmodified
-}
-
-// Create metatable for Color and saves it at LUA_REGISTRYINDEX with name "mt_Color"
-void Lua::createColorMetatable( ){
-
-    luaL_newmetatable( state , "mt_Color" );
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__index" , {
-		lua->pushVariant( arg1.get( arg2 ) );
-		return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__newindex" , {
-		// We can't use arg1 here because we need to reference the userdata
-		((Variant*)lua_touserdata(inner_state,1))->set( arg2 , arg3 );
-		return 0;
-	});	
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__add" , {
-		lua->pushVariant( arg1.operator Color() + arg2.operator Color() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__sub" , {
-		lua->pushVariant( arg1.operator Color() - arg2.operator Color() );
-    return 1;
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__mul" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::COLOR:
-				lua->pushVariant( arg1.operator Color() * arg2.operator Color() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Color() * arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-	LUA_METAMETHOD_TEMPLATE( state , -1 , "__div" , {
-		switch( arg2.get_type() ){
-			case Variant::Type::COLOR:
-				lua->pushVariant( arg1.operator Color() / arg2.operator Color() );
-				return 1;
-			case Variant::Type::INT:
-			case Variant::Type::FLOAT:
-				lua->pushVariant( arg1.operator Color() / arg2.operator double() );
-				return 1;
-			default:
-				return 0;
-		}
-	});
-
-    lua_pop(state,1); // Stack is now unmodified
 }
 
 // Assumes there is a error in the top of the stack. Pops it.
